requirements.txt
typer[all]
pydantic
python-dateutil
beautifulsoup4
lxml
tqdm
playwright
fastapi
uvicorn